move the core dir
