+++ minvqvae/core/__init__.py
@@ -1,0 +1,16 @@
+# quantization
+from .quantize import (
+    # not updated embedding
+    SoftQuantize,
+    SimpleQuantize,
+    # direct updated embedding
+    SoftmaxQuantize,
+    ArgmaxQuantize,
+    # cached update embedding (Implement Later)
+)
+
+# without quantization
+from .ffn import (
+    LLMFFN,
+    FFN,
+)
