+++ minvqvae/core/quantize/__init__.py
@@ -1,7 +1,4 @@
 from .SimpleQuantize import *
-from .VectorQuantize import *
-from .ProductQuantize import *
-from .KmeansQuantize import *
 from .SoftQuantize import *
 from .SoftmaxQuantize import *
 from .ArgmaxQuantize import *
